+++ frontend/app/components/panel-all-sessions-overview.js
@@ -1,10 +1,6 @@
 import Component from '@ember/component';
-import { reads } from '@ember/object/computed';
-
 
 // components/panel-all-sessions-overview.js
 export default Component.extend({
-
-    sessions: null
 
 });
