+++ frontend/app/controllers/session.js
@@ -1,8 +1,12 @@
 import Controller from '@ember/controller';
+import { inject } from '@ember/service';
 import { reads } from '@ember/object/computed';
+
 
 // controllers/session.js
 export default Controller.extend({
+
+    store: inject(),
 
     // Reads `sesssions` from routes/application.js
     sessions: reads('model.sessions')
