+++ frontend/app/router.js
@@ -9,7 +9,6 @@ const Router = EmberRouter.extend({
 Router.map(function() {
   this.route('about');
   this.route('sessions');
-  this.route('application.js');
 });
 
 export default Router;
