+++ frontend/app/routes/sessions.js
@@ -1,10 +1,24 @@
 import Route from '@ember/routing/route';
-import { reads } from '@ember/object/computed';
+import { inject } from '@ember/service';
 
+import RSVP from 'rsvp';
 
 // routes/sessions.js
 export default Route.extend({
 
-    sessions: reads('model.sessions')
+    store: inject(),
+
+    model(/* empty for now */) {
+        const store = this.get('store');
+
+        // Send a call to GET /sessions
+        const sessions = store.findAll('session');
+
+        console.log(sessions);
+        
+        return RSVP.hash({
+            sessions
+        });
+    }
 
 });
