+++ frontend/app/templates/sessions.hbs
@@ -7,8 +7,4 @@
       sessions=sessions
     }}
   </p>
-
-  {{#link-to "about" class="button"}}
-    About Fe
-  {{/link-to}}
 </div>
